perf(bundle): split graph-data.ts color helpers + lazy TimeSeriesOverlay
Pulls the 3000-line graph-data dataset off the critical-path bundle by
extracting the four lightweight helpers (getCategoryColor / getDomainColor
/ getCategoryLabel / EMPTY_GRAPH) into a new graph-color.ts module.
useApexStore — eagerly imported by every page — now only pulls EMPTY_GRAPH
from the new 75-LOC file instead of the whole dataset.

Twelve color-only consumers (CausalDAG2D, CausalDAGMap, NodeInspector,
TarskiPanel, ParetoPanel, DcdGraph, PcmciGraph, FciGraph, DAGOverlay,
DAGNode3D, TimeSeriesOverlay, ClientHeaderBar) repointed to graph-color.
RiskPropagationFlow split its mixed import (getDomainColor → graph-color,
buildRiskCards stays in graph-data). The four MAIN_GRAPH / NODES consumers
keep importing from graph-data; all four are already on lazy paths.

Also dynamic-loads TimeSeriesOverlay (987 LOC) gated on
pinnedTimeSeriesNodes.length > 0 — first paint never has pins so the
chunk defers until the user pins a series.
