+++ src/components/RiskPropagationFlow.tsx
@@ -4,7 +4,8 @@ import { useMemo, useState, useRef, useCallback, useEffect } from "react";
 import { motion, AnimatePresence } from "framer-motion";
 import { useApexStore } from "@/stores/useApexStore";
 import { useFilteredGraph } from "@/hooks/useFilteredGraph";
-import { getDomainColor, buildRiskCards } from "@/lib/graph-data";
+import { getDomainColor } from "@/lib/graph-color";
+import { buildRiskCards } from "@/lib/graph-data";
 import { useTemporalGraph } from "@/hooks/useTemporalGraph";
 import type { NodeTemporalState } from "@/lib/temporal-data";
 import { getNodeDataDescription } from "@/lib/real-timeseries";
