+++ src/components/trinity/FciGraph.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useMemo, useState } from "react";
-import { getCategoryColor } from "@/lib/graph-data";
+import { getCategoryColor } from "@/lib/graph-color";
 import { useApexStore } from "@/stores/useApexStore";
 import { useFilteredGraph } from "@/hooks/useFilteredGraph";
 import { CausalNode } from "@/lib/types";
